Adds 2 new ruins and revamps the icebox version of the tech cult ruin
Adds a abandoned syndicate outpost, which is the syndicate nuke outpost except for being partially destroyed and has a Agent ID to find. Also adds a space ruin which is a ancient nuke op ship. Includes a Deagle and swat armor for aspiring curators to find. The ruin adds 2 simplemobs to fight. Revamps the tech cult ruin, the icemoon version at least, to be bigger and have a more interesting design.