Fix Out of Stock wording on vendors (#1596)
Before the color of the letters "in stock" would indicate if it was in stock or not. Now there is words. If the first vendor creator had a reason for this choice i will reap the unforseen consequences.
The color is still there but now it will say "sold out" or "in stock".