kappa water unsafing (#1858)
water on kappa was having runtimes because i kinda pulled the rug under the mappers when i removed the safe var and replaced it with a proc. So now the water is safe because it has flora ontop of it preventing sinking. I may in the future make all structures prevent sinking... maybe.