Dosage Estimator
A tool for measuring how long the dosage of a substance will last in the human body.
The margin of error between the calculated time and the actual time it took to metabolize the reagent is mostly unnoticable.

Formula:
(volume / metabolization_rate) * 2 = seconds
(seconds/2) * metabolization_rate = volume

add: dosage_estimator tool