Add Visual Flash to Better Memories
Better Memories now camera flashes your screen when they snap a picture of you. This is a purely visual flash that is intended to give memories a chance to escape while also temporarily blinding people.

If this is a good feature or not is yet to be seen so i will atomize this into a compact PR.