Snow Queen's Winter Arrives
Snow Queen is a unique arena based abnormality that is generally weaker outside of her arena.
There is alot of bulky code in this abnormality so i will take any suggestions to reduce the code's impact of preformance even if it only procs once in a while.
add: snow_queen.dm
tweak: abnormality_z map

Add healing zones to arena.