Updates Yin EGO sprites and Fixes sprite misalignments V2 (#2243)
* Updates Yin EGO sprites (Weapon and Armor) and Fixes various sprite misalignments

This is the 2nd attempt as fork did some wacky shit and tried to include a TGS update that already happened. I'm bad at this.

* Updates Yin EGO sprites (Weapon and Armor) and Fixes various sprite misalignments

This is the 2nd attempt as fork did some wacky shit and tried to include a TGS update that already happened. I'm bad at this.