Tweaks Green Stem Ego and New Move_Resist Proc[DONE] (#2312)
Tweaks and Reworks Green Stem Ego

Tweaks and reworks some of the green stem ego code so that it works better as a weapon.

Replaced the throw_at() in /obj/item/ego_weapon/attack() with safe_throw_at() so that move_resist can be used to prevent knockback.

What i did was make a proc that alters the move_resist of the user. move_resist if greater than the force applied by safe_throw_at(force) will cancel the knockback throw but not the damage applied beforehand.

Instead of making it a proc unique to green stem i may move it back to ego_weapons or put the move_resist change proc in mobs.
My other choice was to use anchored but it was a bit wonky since client mobs can still move while anchored and altering a TRUE/FALSE var felt like it'd cause more confusion than nessesary.