Cognitohazard Component and Graffiti [Atomized]
Creates the Cognitohazard component. Its a component that applys a status effect upon being examined. If the obvious setting is turned on then the component will process and afflict humans within a 5 tile observation area with the status effect every 20 seconds.

This is created from butchered manual blinking code.

The graffiti_cognito and the wall markings is for mappers to create cognitohazards in their dungeons.

add: /obj/structure/graffiti_cognito
add: cognitohazard component