Fix Mermaid Water Tiles (#2609)
## About The Pull Request
Fixes mermaid AI. Mermaid on spawn is placed on deep water tiles that are labeled as dense but coded to be non dense in order to force mobs to avoid them. Mermaids AI thinks that she is in a 3by3 wall due to this. I changed obsessing water subtype to nondense. to fix this.
## Changelog
:cl:
tweak: obsessing water is now non dense.
/:cl: