New Adventure Console Events for Teth Abnormalities (#2753)
* Additional Teth Events

More Adventure Console Events for some of the Teth Abnos

* Update beauty_beast.dm

* Finally Fixing Fairy Festival, and Editing

Fairy Festival was fixed, and a umber of the unreleased events were edited.

* Update beauty_beast.dm

* Quick edits to text and variables

* Minor Spelling fix

Fixed a mispelled word in cherry blossom event.

* Update fairy_festival.dm

* Additional Teth Events

More Adventure Console Events for some of the Teth Abnos

* Update beauty_beast.dm

* Finally Fixing Fairy Festival, and Editing

Fairy Festival was fixed, and a umber of the unreleased events were edited.

* Update beauty_beast.dm

* Quick edits to text and variables

* Minor Spelling fix

Fixed a mispelled word in cherry blossom event.

---------

Co-authored-by: Kirie Saito <77302679+Kitsunemitsu@users.noreply.github.com>
Co-authored-by: InsightfulParasite <invaderchris4898@hotmail.com>