Fixes fake shuttle call announcement on shuttle purchase (#2884)
Purchasing a shuttle on LC13, City of Light or City Corp no longer sends a shuttle call announcement

Attempting to purchase a shuttle called the autoEvac() proc in subsystem/shuttle.dm. This PR only adds a check to see if we're on an LC13 map, Corp City or City of Light, and prevents the shuttle from being requested by this proc.