Adds dash attack to lanky sweepers
They get Sweep the Backstreets, once every 12 seconds. They'll try to use it in place of a normal melee attack, it uses the same logic as leader rat but it spawns gibs on hit for visual factor and heals the sweeper a bit. They may or may not become evasive after using it, making them retreat and sidestep to make them hard to catch.