Discord type changed and slight "rebalancing"
"Rebalancing" is in quotes because it seems to me that the previous state was an oversight.

Discord, the weapon, in the current Master branch (6th of June 2025), has one-handed DPS of 40 and wielded DPS of 36. These numbers are relatively low compared to other WAW weapons, which can be explained by their Harmony special feature (healing Assonance weapon users on hit, and being healed by theirs too). However the most glaring issue here is... why wield the weapon if it lowers your DPS?

I have not chosen to rework the weapon or anything of the sort, though I would like to do that. All this PR does is:
- Fixed really minor typo in the description because it bothered me and I couldn't justify not correcting it
- Discord is now a subtype of /ego_weapon/wield/ instead of just /ego_weapon/. It previously implemented its own code for wielding, now it just uses the same framework as Ochre Sheet and Dark Carnival. Yes I remembered to change the datum too, and the reference in Yang's weapon.
- Discord wielded force 20->27. This results in the following DPS numbers (Justice 0):
One handed: 40dps
Wielded: 48.6dps

Justifying this change:
This is in line with other WAW weapons, though I did not find any document outlining intended DPS so I had to go test on a dummy for a while.
From what I see, weapons with helpful gimmicks have lower DPS, and weapons with gimmicks you have to play around have higher DPS. Discord is a weapon with a hard to achieve conditional (you need to talk Agents into actually taking both Yin and Yang, then you need two people to be using Discord and Assonance weapons respectively which is a tough ask honestly), so I think this DPS is reasonable for it. Other numbers from my testing for comparison:
- SSWT (minorly negative gimmick - combo delay, no support aspect) 60dps
- Oppression (minorly negative gimmick - have to correctly time empowering it, no support aspect) around 52dps
- Totalitarianism (negative gimmick - have to wind up each strike by standing still to achieve full potential, no support aspect) around 61dps, 32dps if you do not play around the gimmick
- Blind Rage (kind of negative gimmick, has really good AOE but it hits structures and other Agents, no support aspect) somewhere in the range of 60-70 dps due to DOT variance, has AOE
- Dipsia (positive gimmick, lifesteal) 40dps
- Amrita (spear with self stun, has AOE) 36dps single target, 44dps AOE spam
- Sunyata (normal beatstick with AOE you can simultaneously do) 40dps single target, 66.7dps if you also spam AOE
- Diffraction (no gimmicks) 42dps, which seems kind of low honestly
- Abyssal Route (combo system that can be good for AOE or bad if it gets you into trouble) 39.4dps

Considering all these numbers and their particular quirks, I think 48.6dps for wielding Discord is a very fair number. And if it isn't, just edit the wielded force I guess.